% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_angle.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert Degrees to Radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Numeric. The angle(s) in degrees to convert to radians.}
}
\value{
Numeric. The corresponding angle(s) in radians.
}
\description{
Converts angles from degrees to radians. Supports vectorized input.
}
\examples{
deg2rad(180)
deg2rad(c(0, 90, 180))
}
