% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_cosmology.R
\name{Mpc_to_km}
\alias{Mpc_to_km}
\title{Convert Megaparsecs to Kilometers}
\usage{
Mpc_to_km(mpc)
}
\arguments{
\item{mpc}{A numeric value representing the distance in megaparsecs.}
}
\value{
A numeric value representing the equivalent distance in kilometers (km).
}
\description{
This function converts a distance value from megaparsecs (Mpc) to kilometers (km).
The conversion factor is based on 1 parsec being equivalent to 3.262 light-years,
and 1 light-year being approximately \(9.461 \times 10^{12}\) kilometers.
}
\examples{
Mpc_to_km(1) # Converts 1 Mpc to kilometers
}
