% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_energy_density_fn_T}
\alias{photon_energy_density_fn_T}
\title{Photon Energy Density as a Function of Temperature}
\usage{
photon_energy_density_fn_T(T, unit = "eV")
}
\arguments{
\item{T}{Numeric. Temperature in either eV or Kelvin.}

\item{unit}{Character. The unit of the temperature input. Either
\code{"eV"} (default) or \code{"K"}.}
}
\value{
Numeric. The photon energy density in natural units.
}
\description{
Calculates the photon energy density for a given temperature \eqn{T},
using \eqn{\rho_\gamma = \frac{\pi^2}{15} T^4}.
}
\examples{
photon_energy_density_fn_T(1, "eV")
photon_energy_density_fn_T(300, "K")
}
