% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_energy_density_fn_z}
\alias{photon_energy_density_fn_z}
\title{Photon Energy Density as a Function of Redshift}
\usage{
photon_energy_density_fn_z(z)
}
\arguments{
\item{z}{Numeric. The redshift value.}
}
\value{
Numeric. The photon energy density in natural units.
}
\description{
Calculates the photon energy density at a given cosmological redshift
\eqn{z} by first converting to temperature in eV.
}
\examples{
photon_energy_density_fn_z(1300)
}
