% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_number_density_fn_T}
\alias{photon_number_density_fn_T}
\title{Photon Number Density as a Function of Temperature}
\usage{
photon_number_density_fn_T(temp, unit = "eV")
}
\arguments{
\item{temp}{Numeric. Temperature in either eV or Kelvin.}

\item{unit}{Character. The unit of the temperature input. Either
\code{"eV"} (default) or \code{"K"}.}
}
\value{
Numeric. The photon number density in natural units.
}
\description{
Calculates the photon number density for a given temperature \eqn{T},
using \eqn{n_\gamma = \frac{2\zeta(3)}{\pi^2} T^3}.
}
\examples{
photon_number_density_fn_T(1, "eV")
photon_number_density_fn_T(300, "K")
}
