% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_number_density_fn_z}
\alias{photon_number_density_fn_z}
\title{Photon Number Density as a Function of Redshift}
\usage{
photon_number_density_fn_z(z)
}
\arguments{
\item{z}{Numeric. The redshift value.}
}
\value{
Numeric. The photon number density in natural units.
}
\description{
Calculates the photon number density at a given cosmological redshift
\eqn{z} by first converting to temperature in eV.
}
\examples{
photon_number_density_fn_z(1300)
}
