% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_angle.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert Radians to Degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Numeric. The angle(s) in radians to convert to degrees.}
}
\value{
Numeric. The corresponding angle(s) in degrees.
}
\description{
Converts angles from radians to degrees. Supports vectorized input.
}
\examples{
rad2deg(pi)
rad2deg(c(0, pi / 2, pi))
}
