% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{Saha_Xe}
\alias{Saha_Xe}
\title{Ionization Fraction Using the Saha Equation}
\usage{
Saha_Xe(z)
}
\arguments{
\item{z}{Numeric. The redshift value.}
}
\value{
Numeric. The ionization fraction \eqn{X_e} (between 0 and 1).
}
\description{
Calculates the ionization fraction \eqn{X_e} as a function of redshift
\eqn{z} using the Saha equation for hydrogen recombination.
}
\examples{
Saha_Xe(1300)
Saha_Xe(1100)
}
