% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_constants.R
\name{constant_value}
\alias{constant_value}
\title{Retrieve the Value of a Physical Constant}
\usage{
constant_value(constant_name, unit = "SI")
}
\arguments{
\item{constant_name}{Character. The name (or part of the name) of the constant to search for.
Case-insensitive and partial matches are allowed.}

\item{unit}{Character. The unit system to retrieve the constant in. Options are:
\itemize{
  \item `"SI"`: Retrieves the value in SI units (default).
  \item `"Natural"`: Retrieves the value in natural units.
}}
}
\value{
A list containing:
  \itemize{
    \item `name`: The name of the constant.
    \item `value`: The value of the constant in the selected unit system.
    \item `unit`: The unit of the constant in the selected unit system.
  }
}
\description{
This function retrieves the value and unit of a specified physical constant from the `constants_df` dataset.
You can search by the constant's name and choose whether to retrieve the value in SI units or natural units.
}
\examples{
# Retrieve the value of the speed of light in SI units
constant_value("speed of light", unit = "SI")
}
