% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_angle.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert Degrees to Radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Numeric. The angle(s) in degrees to convert to radians.}
}
\value{
Numeric. The corresponding angle(s) in radians.
}
\description{
This function converts angles from degrees to radians. It supports vectorized input.
}
\examples{
deg2rad(180) # Output: 3.141593 (pi)
deg2rad(c(0, 90)) # Output: 0, 1.570796
}
