% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_angle.R
\name{deg_to_hms}
\alias{deg_to_hms}
\title{Position and Motion of Stars Conversion Functions}
\usage{
deg_to_hms(deg, type = "cat", digit = 5)
}
\arguments{
\item{deg}{[ADD DESCRIPTION]}

\item{type}{[ADD DESCRIPTION]}

\item{digit}{[ADD DESCRIPTION]}
}
\description{
Normally in astronomy, different angular systems are used for location of a given planet, star or other phenomena. To deal with different angular systems and their conversions a suite of conversion functions are employed. `deg_to_hms()` ...[TODO]
}
