% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gaia.R
\name{get_gaia_data}
\alias{get_gaia_data}
\title{Query Gaia Archive Data}
\usage{
get_gaia_data(vars, condition)
}
\arguments{
\item{vars}{A character string specifying the variables to retrieve
(e.g., "source_id, ra, dec").}

\item{condition}{A character string specifying the conditions to filter the data
(e.g., "parallax > 10").}
}
\value{
A data frame containing the requested data.
}
\description{
This function queries the Gaia Archive TAP service to retrieve data based on
the specified variables and conditions.
}
\details{
It uses the Gaia Early Data Release 3 (EDR3) catalog, which contains information

This function sends a synchronous query to the Gaia Archive TAP service
}
\note{
This function requires the `httr` and `jsonlite` packages to handle HTTP
requests and parse JSON responses, respectively.
}
\examples{
# Define the variables and condition
vars <- "source_id, ra, dec, phot_bp_mean_mag, phot_rp_mean_mag, parallax"
condition <- "parallax > 40"
# Fetch the data from the Gaia Archive
result <- get_gaia_data(vars, condition)
head(result)

}
