% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{hms_to_deg}
\alias{hms_to_deg}
\title{Position and Motion of Stars Conversion Functions}
\usage{
hms_to_deg(h, m, s, digit = 5)
}
\arguments{
\item{h}{[ADD DESCRIPTION]}

\item{m}{[ADD DESCRIPTION]}

\item{s}{[ADD DESCRIPTION]}

\item{digit}{[ADD DESCRIPTION]}
}
\description{
Normally in astronomy, different angular systems are used for location of a given planet, star or other phenomena. To deal with different angular systems and their conversions a suite of conversion functions are employed. `hms_to_deg()` ...[TODO]
}
