% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_energy_density_fn_T}
\alias{photon_energy_density_fn_T}
\title{Photon Energy Density as a Function of Temperature}
\usage{
photon_energy_density_fn_T(T, unit = "eV")
}
\arguments{
\item{T}{Numeric. Temperature in either eV or Kelvin.}

\item{unit}{Character. The unit of temperature, either "eV" or "K". Defaults to "eV".}
}
\value{
Numeric. The photon energy density.
}
\description{
This function calculates the photon energy density for a given temperature \(T\).
}
\examples{
photon_energy_density_fn_T(1, "eV")
photon_energy_density_fn_T(300, "K")
}
