% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{photon_number_density_fn_z}
\alias{photon_number_density_fn_z}
\title{Photon Number Density as a Function of Redshift}
\usage{
photon_number_density_fn_z(z)
}
\arguments{
\item{z}{Numeric. The redshift value.}
}
\value{
Numeric. The photon number density.
}
\description{
This function calculates the photon number density for a given redshift \(z\).
}
\examples{
photon_number_density_fn_z(1300)
}
