% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_angle.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert Radians to Degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Numeric. The angle(s) in radians to convert to degrees.}
}
\value{
Numeric. The corresponding angle(s) in degrees.
}
\description{
This function converts angles from radians to degrees. It supports vectorized input.
}
\examples{
rad2deg(pi) # Output: 180
rad2deg(c(0, pi / 2)) # Output: 0, 90
}
