% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_thermo2.R
\name{Saha_Xe}
\alias{Saha_Xe}
\title{Ionization Fraction Using the Saha Equation}
\usage{
Saha_Xe(z)
}
\arguments{
\item{z}{Numeric. The redshift value.}
}
\value{
Numeric. The ionization fraction \(X_e\).
}
\description{
This function calculates the ionization fraction \(X_e\) as a function of redshift \(z\) using the Saha equation.
}
\examples{
Saha_Xe(1300)
}
